Update table template - Add delete confirm

+++ views/partials/table.tpl
@@ -16,7 +16,7 @@
 			{/foreach}
 			<td class="center">
 				<a href="{$table->edit_url}/{$data->id}"><i class="fa fa-pencil fa-fw"></i></a>
-				<a href="{$table->delete_url}/{$data->id}"><i class="fa fa-trash fa-fw"></i></a>
+				<a href="{$table->delete_url}/{$data->id}" onclick="if(confirm('{t}Are you sure you want to delete this{/t} {$table->entity} ({$data->id}) ?')) { return true } return false"><i class="fa fa-trash fa-fw"></i></a>
 			</td>
 		</tr>
 		{assign var=i value=$i + 1}
